+++ src/shared/components/dropdown-deprecated/index.ts
@@ -1,0 +1,1 @@
+export * from './dropdown'
