+++ src/shared/components/dropdown/dropdown.tsx
@@ -1,176 +1,105 @@
 'use client'
 
-import { useRef, useState } from 'react'
+import { cloneElement, isValidElement, useRef } from 'react'
+import { ReactElement, ReactNode } from 'react'
 
-import styled from '@emotion/styled'
+import { useOutsideClick, useToggle } from '@/shared/hooks'
 
-import { LucideIcon } from '@/shared/components'
-import { theme } from '@/shared/config'
-import { useKeyClose, useOutsideClick, useToggle } from '@/shared/hooks'
+import { DropdownContext, useDropdownContext } from './dropdown-context'
+import * as S from './style'
 
-export interface DropdownProps<OptionType> {
-  label: string
-  getOptionLabel: (option: OptionType) => string
-
-  options: OptionType[]
-  getOptionKey: (option: OptionType) => string | number
-  selected?: OptionType
-
-  onChange: (value: OptionType) => void
-
-  maxVisibleCount?: number
-  className?: string
-  required?: boolean
+interface DropdownProps<T> {
+  children: ReactNode
+  selected: T | undefined // 아직 선택되지 않은 경우 undefined
+  onSelect: (value: T) => void
 }
 
-export const Dropdown = <OptionType,>({
-  label,
-  options,
-  selected,
-  onChange,
-  getOptionLabel,
-  getOptionKey,
-  maxVisibleCount = 4,
-  className,
-  required,
-}: DropdownProps<OptionType>) => {
-  const { value: isOpen, toggle, setValue: setIsOpen } = useToggle(false)
-  const dropdownRef = useRef<HTMLUListElement>(null)
+export const Dropdown = <T,>({ children, selected, onSelect }: DropdownProps<T>) => {
+  const { value: isOpen, toggle, setValue } = useToggle(false)
 
-  const [hoveredIndex, setHoveredIndex] = useState<number | null>(null)
-
-  const handleSelect = (value: OptionType) => {
-    onChange(value)
-    setIsOpen(false)
-  }
-
-  useOutsideClick(dropdownRef as React.RefObject<HTMLElement>, toggle)
-  useKeyClose('Escape', dropdownRef as React.RefObject<HTMLElement>, toggle)
+  // 외부 클릭시 닫기
+  const ref = useRef<HTMLDivElement>(null)
+  useOutsideClick(ref as React.RefObject<HTMLElement>, () => setValue(false))
 
   return (
-    <Wrapper className={className}>
-      <Label isFocused={isOpen || !!selected}>
-        {label}
-        {required && <RequiredMark>*</RequiredMark>}
-      </Label>
-      <SelectBox
-        onClick={toggle}
-        isFocused={isOpen || !!selected}
-        role='combobox'
-        aria-expanded={isOpen}
-        aria-required={required}
-        tabIndex={0}
-      >
-        <SelectedText>{selected ? getOptionLabel(selected) : ''}</SelectedText>
-        <IconWrapper isFocused={isOpen}>
-          <LucideIcon name='ChevronDown' size={16} />
-        </IconWrapper>
-      </SelectBox>
-      {isOpen && (
-        <DropdownBox ref={dropdownRef} maxHeight={maxVisibleCount * 40}>
-          {options.map((option, index) => (
-            <Item
-              key={getOptionKey(option)}
-              onClick={() => handleSelect(option)}
-              onMouseEnter={() => setHoveredIndex(index)}
-              onMouseLeave={() => setHoveredIndex(null)}
-              isHovered={hoveredIndex === index}
-            >
-              {getOptionLabel(option)}
-            </Item>
-          ))}
-        </DropdownBox>
-      )}
-    </Wrapper>
+    <DropdownContext.Provider
+      value={{
+        selected,
+        onSelect,
+        isOpen,
+        toggle,
+      }}
+    >
+      <S.Wrapper ref={ref}>{children}</S.Wrapper>
+    </DropdownContext.Provider>
   )
 }
 
-// === Styles ===
+interface TriggerProps {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  as: ReactElement<any>
+}
 
-const Wrapper = styled.div`
-  width: 200px;
-  position: relative;
-`
+/**
+ * 드롭다운 트리거 요소
+ * @param as 드롭다운 트리거 컴포넌트
+ */
+const Trigger = ({ as }: TriggerProps) => {
+  const { toggle } = useDropdownContext()
 
-const Label = styled.label<{ isFocused: boolean }>`
-  position: absolute;
-  top: ${({ isFocused }) => (isFocused ? '-12px' : '50%')};
-  transform: translateY(${({ isFocused }) => (isFocused ? '0' : '-50%')});
+  if (!isValidElement(as)) return null
 
-  font-size: ${({ isFocused }) => (isFocused ? theme.fontSize.xs : theme.fontSize.sm)};
-  font-weight: ${theme.fontWeight.medium};
+  const child = as as ReactElement<{ onClick?: (e: React.MouseEvent<HTMLElement>) => void }>
 
-  color: ${theme.colors.lfBlack.base};
-  transition: all 0.2s ease;
-  pointer-events: none;
-`
+  return cloneElement(child, {
+    onClick: (e: React.MouseEvent<HTMLElement>) => {
+      child.props.onClick?.(e)
+      toggle()
+    },
+  })
+}
 
-const RequiredMark = styled.span`
-  color: ${theme.colors.lfGreenBorder.base};
-  margin-left: 4px;
-`
+interface MenuProps {
+  children: ReactNode
+  maxVisibleCount?: number
+}
 
-const SelectBox = styled.div<{ isFocused: boolean }>`
-  padding: 6px 8px;
-  border: none;
-  border-bottom: 2px solid ${theme.colors.lfLightGray.base};
-  background: transparent;
-  cursor: pointer;
-  position: relative;
+/**
+ * 열린 드롭다운의 영역
+ */
+const Menu = ({ children, maxVisibleCount = 4 }: MenuProps) => {
+  const { isOpen } = useDropdownContext()
+  if (!isOpen) return null
 
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+  return <S.MenuWrapper maxHeight={maxVisibleCount * 40}>{children}</S.MenuWrapper>
+}
 
-  &::before {
-    content: '';
-    position: absolute;
-    bottom: -2px;
-    left: 50%;
-    transform: translateX(-50%) scaleX(${({ isFocused }) => (isFocused ? 1 : 0)});
-    transform-origin: center;
-    width: 100%;
-    height: 2px;
-    background-color: ${theme.colors.lfBlack.base};
-    transition: transform 0.3s ease;
+interface ItemProps<T> {
+  value: T
+  children: ReactNode
+}
+
+/**
+ * 하나의 드롭다운
+ */
+const Item = <T,>({ value, children }: ItemProps<T>) => {
+  const { onSelect, toggle } = useDropdownContext()
+
+  const handleClick = () => {
+    onSelect(value)
+    toggle()
   }
-`
 
-const SelectedText = styled.span`
-  font-size: ${theme.fontSize.xs};
-  font-weight: ${theme.fontWeight.medium};
-  color: ${theme.colors.lfBlack.base};
-`
+  const child = children as ReactElement<{ onClick?: (e: React.MouseEvent<HTMLElement>) => void }>
 
-const IconWrapper = styled.div<{ isFocused: boolean }>`
-  transform: rotate(${({ isFocused }) => (isFocused ? 180 : 0)}deg);
-  transition: transform 0.3s ease;
-`
+  return cloneElement(child, {
+    onClick: (e: React.MouseEvent<HTMLElement>) => {
+      child.props.onClick?.(e)
+      handleClick()
+    },
+  })
+}
 
-const DropdownBox = styled.ul<{ maxHeight: number }>`
-  position: absolute;
-  top: 100%;
-  left: 0;
-  width: 100%;
-  max-height: ${({ maxHeight }) => maxHeight}px;
-  overflow-y: auto;
-  margin-top: 8px;
-  padding: 8px 4px;
-  border-radius: ${theme.radius.base};
-  box-shadow: ${theme.shadow.lfInput};
-  background-color: ${theme.colors.lfWhite.base};
-  z-index: 10;
-  list-style: none;
-`
-
-const Item = styled.li<{ isHovered: boolean }>`
-  padding: 12px 4px;
-  text-align: center;
-  font-size: ${theme.fontSize.xs};
-  font-weight: ${theme.fontWeight.regular};
-  color: ${theme.colors.lfBlack.base};
-  background-color: ${({ isHovered }) => (isHovered ? theme.colors.lfLightGray.base : 'transparent')};
-  cursor: pointer;
-  border-radius: ${theme.radius.sm};
-  transition: background-color 0.2s ease;
-`
+Dropdown.Trigger = Trigger
+Dropdown.Menu = Menu
+Dropdown.Item = Item
