+++ src/shared/components/index.ts
@@ -5,6 +5,7 @@ export * from './calender'
 export * from './check-icon'
 export * from './datepicker'
 export * from './dropdown'
+export * from './dropdown-deprecated'
 export * from './errortext'
 export * from './header'
 export * from './image-input'
@@ -14,6 +15,7 @@ export * from './leaf-reward'
 export * from './loading'
 export * from './modal'
 export * from './navbar'
+export * from './select'
 export * from './slidearea'
 export * from './switchtap'
 export * from './toast'
